Modified grid-accell to keep a set of PokeRays rather than a set of rays. A PokeRay is nothing else than a ray which is allowed to "poke" only one specific face.
I changed the class GridAccell to allow to keep track of pokerays rather
than simple rays and modified "mode == FACEINTERSECTIONS" in Balloon::updateViewField
to use these structures instead. Note that when using this mode you discard rays
which have t<0, so you don't even need to re-specify anything in the accellerator, 
as wahtever will be used by the next mode will be simply discarded...

I also added a new mode (STILL IN DEVELOPMENT) called BIFACEINTERSECTIONS which 
do not only look at the intersections found by a ray in the +side but also 
takes a look behind the scanpoint.