change install meshlabserver shortcut
TO BE TESTED
we have to avoid to overwrite the existing "cmd" links. Since the link target is now "cmd.exe /k $INSTDIR\meshlabserver.exe", it should be seen as a different target, and hopefully not overwrite.
The effect is the same, except that meshlabserver is called once, so you can see the help, which is not a bad idea :)