qPrintable to qUtf8Printable in the WHOLE project
qUtf8Printable is better to deal with special characters.
also in QT documentation it is said is better when logging strings.
However, we use it also to store filenames and text-based ID, if we only use one, and not alternating between qUtf8Printable and qPrintable, thins should be smoother, but since it is a project-wide change, it is potentially DANGEROUS, NOT YET FULLY TESTED