Put Start shortcuts in allusers profile instead of currentuser
The default shell context in NSIS is for the current user. This causes $SMPROGRAMS to put the link in the current user's profile. Other users on the system get no start menu link. This is causing me trouble because I'm trying to deploy Meshlab in a school and by default only the user that installs the application gets the Start Menu shortcuts.

Docs for SetShellVarContext at https://nsis.sourceforge.io/Docs/Chapter4.html