Fix depthmap shader coloring
I'm brand new here, but I was trying to use the depthmap.gdp shader and
wasn't seeing the results I expected. Poking around the fragment shader,
I noticed the `gray` value was used for red, green, blue *and* alpha
channels, meaning values towards 0 (black) won't be displayed as
expected. My wild guess is that this worked when the alpha channel was
ignored, but for me on MacOS it was definitely not working.

Using the `gray` value for red/green/blue and 1.0 for alpha fixes it.