Install app icon in XDG hicolor icon theme
Install the icon of the application in the hicolor XDG icon theme;
this way it can be properly loaded by XDG menus in the currently set
XDG icon theme, without looking in the legacy pixmaps location.

As result, adapt the desktop file by removing the full path to the icon
(which was already buggy for installations in prefixes different than
/usr), so the bare icon name will be searched in the XDG icon theme as
expected.

Also, remove the manual icon copying in a bundle script, as the icon is
installed in the right place now.