Consumes the values of MLMatrix44 one by one.
The MLMatrix44 element is of type of the Vector16 simple type
that is a restriction of the other VectorList simple type
that is a list of decimal simple type.
MeshLabProject XML schema: meshlab/docs/meshlabprojspecificationformat.xsd

So values are separed by spaces (by XML list definition) and
carriage return are ignored (by XML definition).

Consequently, there is no need to prejudge the presence and position
of any line breaks.